pyproject.toml
[tool.poetry]
name = "bid"
version = "0.1.0"
description = ""
authors = ["Rafael <rafael.moreno@agileengine.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10 <4.0"
django = "^5.0.6"
django-rest-framework = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
